commandCompletionTimer variable now changes every time 3 buttons are pressed correctly

+++ Glossalalia/Glossalalia/GALevelViewController.h
@@ -32,7 +32,7 @@
 @property int score;
 @property int numWordsCorrect;
 @property int numButtonWordsPerPlayer;
-@property float commandCompletionTimeLimit;
+@property CGFloat commandCompletionTimeLimit;
 @property float commandCompletionTimeRemaining;
 
 @property NSMutableArray *legalCommandWords;
